Draw a calendar grid
The grid can either have 5 or 6 rows, depending on the month. There is not code
yet for determining the number of days in a month, or the weekday the month
begins with. But it works if you set those manually.

Threw away the code that shows the date for now.