Use the correct number of days in the month, and add date selection
The highlighted date can now be changed by using the arrow keys. It is not yet
possible to go to another month - the selection then just disappears. Also the
name of the current month is not yet shown.

Also, fixed a stupid bug in weekdayYearStart, where the algorithm only worked
correctly when b was zero, and thus failed in subsequent invocations.