Start working on the day view
The code for drawing the day view is located in the new file
dayview.asm. Furthermore, the code for drawing the month view is split
off to the new file monthview.asm. We can now switch between the views
by simply doing kjp(monthView) and kjp(dayView).

The day view does not do anything yet, except for displaying an empty
window.

Also, introduced a menu. Currently the menu only displays an option to
quit the application, but in the day view eventually options will be
added to add appointments and so on.