Fix bug in the initial month data calculation
When the program starts, it computes and stores the ‘month data’ of the
current month (number of days and the weekday the 1st day of the month).
This month data is then recomputed whenever the user moves to a
different month.

Unfortunately the initial computation of the month data was incorrect
due to the parameters of updateMonthData not being set properly. This
commit ensures those parameters are set.